Implement placeholder structure for updated about page
